Fix everything up, kinda usable again

+++ Trakr/NSString+URLEncode.m
@@ -13,8 +13,8 @@
 - (NSString *) URLEncodedString {
     NSMutableString * output = [NSMutableString string];
     const unsigned char * source = (const unsigned char *)[self UTF8String];
-    int sourceLen = strlen((const char *)source);
-    for (int i = 0; i < sourceLen; ++i) {
+    unsigned long sourceLen = strlen((const char *)source);
+    for (unsigned int i = 0; i < sourceLen; ++i) {
         const unsigned char thisChar = source[i];
         if (thisChar == ' '){
             [output appendString:@"+"];
